Apport de plus de sécurité contre les injections pour les reqûtes SQL

+++ server/controllers/posts.js
@@ -16,7 +16,8 @@ exports.getAllPosts = (req, res, next) => {
 //Gets one post of the Data Base thanks to its ID
 exports.getOnePost = (req, res, next) => {
   connection.query(
-    `SELECT * FROM posts WHERE id_posts = ${req.params.id}`,
+    `SELECT * FROM posts WHERE id_posts = ?`,
+    [req.params.id],
     function (error, results, fields) {
       if (error) {
         res.status(500).json({ error });
@@ -38,7 +39,8 @@ exports.createOnePost = (req, res, next) => {
       }
     : { ...req.body };
   connection.query(
-    `INSERT INTO posts (content, image_url, users_id_users) VALUES ('${postObject.content}' , ${postObject.imageUrl} , ${req.auth.userId})`,
+    `INSERT INTO posts (content, image_url, users_id_users) VALUES (? , ? , ?)`,
+    [postObject.content, postObject.imageUrl, eq.auth.userId],
     function (error, results, fields) {
       if (error) {
         res.status(500).json({ error });
@@ -63,27 +65,29 @@ exports.modifyOnePost = (req, res, next) => {
 // Deletes one post of the Data Base thanks to its ID
 exports.deleteOnePost = (req, res, next) => {
   connection.query(
-    `SELECT * FROM posts WHERE id_posts = ${req.params.id}`,
+    `SELECT * FROM posts WHERE id_posts = ?`,
+    [req.params.id],
     function (error, results, fields) {
-      //   if (results[0].users_id_users != req.auth.userId) {
-      //     res.status(403).json({ message: 'Unauthorized request' });
-      //   } else {
-      if (results[0].image_url != null) {
-        const filename = results[0].image_url.split('/images/')[1];
-        fs.unlink(`images/${filename}`, () => {});
-      }
-      connection.query(
-        `DELETE FROM posts WHERE id_posts = ${req.params.id}`,
-        function (error, results, fields) {
-          if (error) {
-            res.status(500).json({ error });
-          } else {
-            res.status(200).json({ message: 'post supprimé !' });
-          }
+      if (results[0].users_id_users != req.auth.userId) {
+        res.status(403).json({ message: 'Unauthorized request' });
+      } else {
+        if (results[0].image_url != null) {
+          const filename = results[0].image_url.split('/images/')[1];
+          fs.unlink(`images/${filename}`, () => {});
         }
-      );
+        connection.query(
+          `DELETE FROM posts WHERE id_posts = ?`,
+          [req.params.id],
+          function (error, results, fields) {
+            if (error) {
+              res.status(500).json({ error });
+            } else {
+              res.status(200).json({ message: 'post supprimé !' });
+            }
+          }
+        );
+      }
     }
-    // }
   );
 };
 
@@ -91,7 +95,8 @@ exports.deleteOnePost = (req, res, next) => {
 exports.manageLike = (req, res, next) => {
   if (req.body.like == 1) {
     connection.query(
-      `INSERT INTO appreciate (users_id_users, posts_id_posts) VALUES (${req.body.userId}, ${req.params.id})`,
+      `INSERT INTO appreciate (users_id_users, posts_id_posts) VALUES (?, ?)`,
+      [req.body.userId, req.params.id],
       function (error, results, fields) {
         if (error) {
           res.status(500).json({ error });
@@ -102,7 +107,8 @@ exports.manageLike = (req, res, next) => {
     );
   } else if (req.body.like == 0) {
     connection.query(
-      `DELETE FROM appreciate WHERE users_id_users = ${req.body.userId} AND posts_id_posts = ${req.params.id}`,
+      `DELETE FROM appreciate WHERE users_id_users = ? AND posts_id_posts = ?`,
+      [req.body.userId, req.params.id],
       function (error, resutls, fields) {
         if (error) {
           res.status(500).json({ error });
